Adds wcsatt metadata to order item and enables resubscribe
Fixes #68

The cart data is very different when resubscribing to a subscription than when we're just checking out. Because of that previously resubscribes were not possible as STATT wasn't able to recreate the data necessary to turn a simple product into a subscription.

This PR allows STATT to store the scheme ID against the order item as an order item meta, thus making it easy to retrieve it when customer resubscribes. It's needed because the scheme data cannot be accurately reconstructed from previous subscription's data and product's available schemes as the length of the subscription is not stored on the subscription meta.

To test:

* purchase a statt product as a subscription
* cancel the subscription
* resubscribe to the subscription

Without this PR: cart and new order will be one off, no subscription will be created.

With this PR: cart has recurring pricing, order is created, and subscription is also created.