Leverge scheme auto-switching in WCS_ATT_Product class
This rounds off the WCS_ATT_Product class based on the discussions we had in #42 -- most of these functions can now be called with a `$scheme_key` argument, which makes it easy to get the property we want and let the API handle scheme switching (if we are asking for something in a different scheme than the one that's currently set on the object).

The next commit(s) will make this concept clearer.