Add global option to control "Add to subscription" behavior
closes #294

Also does away with the 'subscription_management_add_to_matching_subscription' feature, as it seems off-place to my current version of self. The 'subscription_management_add_to_matching_subscription' feature can't be turned off anyway and was used in the code as a shorter way to tell if a product (for which add-to-sub is supported) has subscription schemes. On seeing it, I initially thought it could be used to disable the "matching" process for products with schemes, but that's not what my old self had in mind with this. Looking closer at all parts of the code where it's used, it became clear that it wasn't actually needed.